remove unused code, reversing parameters
