+++ clients/include/testing_nrm2_batched.hpp
@@ -35,12 +35,12 @@ void testing_nrm2_batched_bad_arg_template(const Arguments& arg)
     CHECK_ROCBLAS_ERROR(rocblas_set_pointer_mode(handle, rocblas_pointer_mode_device));
 
     EXPECT_ROCBLAS_STATUS(
-        (rocblas_nrm2_batched<T1, T2>(handle, N, nullptr, incx, d_rocblas_result, batch_count)),
+        (rocblas_nrm2_batched<T1, T2>(handle, N, nullptr, incx, batch_count, d_rocblas_result)),
         rocblas_status_invalid_pointer);
-    EXPECT_ROCBLAS_STATUS((rocblas_nrm2_batched<T1, T2>(handle, N, dx, incx, nullptr, batch_count)),
+    EXPECT_ROCBLAS_STATUS((rocblas_nrm2_batched<T1, T2>(handle, N, dx, incx, batch_count, nullptr)),
                           rocblas_status_invalid_pointer);
     EXPECT_ROCBLAS_STATUS(
-        (rocblas_nrm2_batched<T1, T2>(nullptr, N, dx, incx, d_rocblas_result, batch_count)),
+        (rocblas_nrm2_batched<T1, T2>(nullptr, N, dx, incx, batch_count, d_rocblas_result)),
         rocblas_status_invalid_handle);
 }
 
@@ -70,7 +70,7 @@ void testing_nrm2_batched_template(const Arguments& arg)
 
         CHECK_ROCBLAS_ERROR(rocblas_set_pointer_mode(handle, rocblas_pointer_mode_host));
         EXPECT_ROCBLAS_STATUS(
-            (rocblas_nrm2_batched<T1, T2>(handle, N, dx, incx, d_rocblas_result, batch_count)),
+            (rocblas_nrm2_batched<T1, T2>(handle, N, dx, incx, batch_count, d_rocblas_result)),
             rocblas_status_invalid_size);
         return;
     }
@@ -89,7 +89,7 @@ void testing_nrm2_batched_template(const Arguments& arg)
 
         CHECK_ROCBLAS_ERROR(rocblas_set_pointer_mode(handle, rocblas_pointer_mode_device));
         CHECK_ROCBLAS_ERROR(
-            (rocblas_nrm2_batched<T1, T2>(handle, N, dx, incx, d_rocblas_result, batch_count)));
+            (rocblas_nrm2_batched<T1, T2>(handle, N, dx, incx, batch_count, d_rocblas_result)));
         return;
     }
 
@@ -143,12 +143,12 @@ void testing_nrm2_batched_template(const Arguments& arg)
         // GPU BLAS, rocblas_pointer_mode_host
         CHECK_ROCBLAS_ERROR(rocblas_set_pointer_mode(handle, rocblas_pointer_mode_host));
         CHECK_ROCBLAS_ERROR((
-            rocblas_nrm2_batched<T1, T2>(handle, N, dx_pvec, incx, rocblas_result_1, batch_count)));
+            rocblas_nrm2_batched<T1, T2>(handle, N, dx_pvec, incx, batch_count, rocblas_result_1)));
 
         // GPU BLAS, rocblas_pointer_mode_device
         CHECK_ROCBLAS_ERROR(rocblas_set_pointer_mode(handle, rocblas_pointer_mode_device));
         CHECK_ROCBLAS_ERROR((rocblas_nrm2_batched<T1, T2>(
-            handle, N, dx_pvec, incx, d_rocblas_result_2, batch_count)));
+            handle, N, dx_pvec, incx, batch_count, d_rocblas_result_2)));
         CHECK_HIP_ERROR(hipMemcpy(
             rocblas_result_2, d_rocblas_result_2, batch_count * sizeof(T2), hipMemcpyDeviceToHost));
 
@@ -192,14 +192,14 @@ void testing_nrm2_batched_template(const Arguments& arg)
 
         for(int iter = 0; iter < number_cold_calls; iter++)
         {
-            rocblas_nrm2_batched<T1, T2>(handle, N, dx_pvec, incx, rocblas_result_2, batch_count);
+            rocblas_nrm2_batched<T1, T2>(handle, N, dx_pvec, incx, batch_count, rocblas_result_2);
         }
 
         gpu_time_used = get_time_us(); // in microseconds
 
         for(int iter = 0; iter < number_hot_calls; iter++)
         {
-            rocblas_nrm2_batched<T1, T2>(handle, N, dx_pvec, incx, rocblas_result_2, batch_count);
+            rocblas_nrm2_batched<T1, T2>(handle, N, dx_pvec, incx, batch_count, rocblas_result_2);
         }
 
         gpu_time_used = (get_time_us() - gpu_time_used) / number_hot_calls;
