+++ library/include/rocblas-functions.h
@@ -583,42 +583,43 @@ ROCBLAS_EXPORT rocblas_status rocblas_dznrm2(rocblas_handle                handl
     @param[in]
     incx      rocblas_int
               specifies the increment for the elements of each x_i. incx must be > 0.
+    @param[in]
+    batch_count rocblas_int
+              number of instances in the batch
     @param[out]
     result
               pointers to array of batch_count size for nrm2 results. either on the host CPU or device GPU.
               return is 0.0 if n <= 0, incx<=0.
-    @param[in]
-    batch_count rocblas_int
-              number of instances in the batch
+
     ********************************************************************/
 
 ROCBLAS_EXPORT rocblas_status rocblas_snrm2_batched(rocblas_handle     handle,
                                                     rocblas_int        n,
                                                     const float* const x[],
                                                     rocblas_int        incx,
-                                                    float*             results,
-                                                    rocblas_int        batch_count);
+                                                    rocblas_int        batch_count,
+                                                    float*             results);
 
 ROCBLAS_EXPORT rocblas_status rocblas_dnrm2_batched(rocblas_handle      handle,
                                                     rocblas_int         n,
                                                     const double* const x[],
                                                     rocblas_int         incx,
-                                                    double*             results,
-                                                    rocblas_int         batch_count);
+                                                    rocblas_int         batch_count,
+                                                    double*             results);
 
 ROCBLAS_EXPORT rocblas_status rocblas_scnrm2_batched(rocblas_handle                     handle,
                                                      rocblas_int                        n,
                                                      const rocblas_float_complex* const x[],
                                                      rocblas_int                        incx,
-                                                     float*                             results,
-                                                     rocblas_int batch_count);
+                                                     rocblas_int                        batch_count,
+                                                     float*                             results);
 
 ROCBLAS_EXPORT rocblas_status rocblas_dznrm2_batched(rocblas_handle                      handle,
                                                      rocblas_int                         n,
                                                      const rocblas_double_complex* const x[],
                                                      rocblas_int                         incx,
-                                                     double*                             results,
-                                                     rocblas_int batch_count);
+                                                     rocblas_int batch_count,
+                                                     double*     results);
 
 /*! \brief BLAS Level 1 API
 
@@ -640,13 +641,14 @@ ROCBLAS_EXPORT rocblas_status rocblas_dznrm2_batched(rocblas_handle             
               specifies the increment for the elements of each x_i. incx must be > 0.
     @param[in]
     stridex   specifies the pointer increment between batches for x. stridex must be >= n*incx
+    @param[in]
+    batch_count rocblas_int
+              number of instances in the batch
     @param[out]
     results
               pointer to array for storing contiguous batch_count results. either on the host CPU or device GPU.
               return is 0.0 if n <= 0, incx<=0.
-    @param[in]
-    batch_count rocblas_int
-              number of instances in the batch
+
     ********************************************************************/
 
 ROCBLAS_EXPORT rocblas_status rocblas_snrm2_strided_batched(rocblas_handle handle,
@@ -654,32 +656,32 @@ ROCBLAS_EXPORT rocblas_status rocblas_snrm2_strided_batched(rocblas_handle handl
                                                             const float*   x,
                                                             rocblas_int    incx,
                                                             rocblas_int    stridex,
-                                                            float*         results,
-                                                            rocblas_int    batch_count);
+                                                            rocblas_int    batch_count,
+                                                            float*         results);
 
 ROCBLAS_EXPORT rocblas_status rocblas_dnrm2_strided_batched(rocblas_handle handle,
                                                             rocblas_int    n,
                                                             const double*  x,
                                                             rocblas_int    incx,
                                                             rocblas_int    stridex,
-                                                            double*        results,
-                                                            rocblas_int    batch_count);
+                                                            rocblas_int    batch_count,
+                                                            double*        results);
 
 ROCBLAS_EXPORT rocblas_status rocblas_scnrm2_strided_batched(rocblas_handle               handle,
                                                              rocblas_int                  n,
                                                              const rocblas_float_complex* x,
                                                              rocblas_int                  incx,
                                                              rocblas_int                  stridex,
-                                                             float*                       results,
-                                                             rocblas_int batch_count);
+                                                             rocblas_int batch_count,
+                                                             float*      results);
 
 ROCBLAS_EXPORT rocblas_status rocblas_dznrm2_strided_batched(rocblas_handle                handle,
                                                              rocblas_int                   n,
                                                              const rocblas_double_complex* x,
                                                              rocblas_int                   incx,
                                                              rocblas_int                   stridex,
-                                                             double*                       results,
-                                                             rocblas_int batch_count);
+                                                             rocblas_int batch_count,
+                                                             double*     results);
 
 /*! \brief BLAS Level 1 API
