+++ library/src/blas1/rocblas_asum_strided_batched.cpp
@@ -70,8 +70,7 @@ namespace
         if(batch_count < 0 || stridex < 0 || stridex < n * incx) // negative n
             return rocblas_status_invalid_size;
 
-        size_t dev_bytes
-            = rocblas_reduction_kernel_workspace_size<NB>(n, batch_count, results);
+        size_t dev_bytes = rocblas_reduction_kernel_workspace_size<NB>(n, batch_count, results);
 
         if(handle->is_device_memory_size_query())
             return handle->set_optimal_device_memory_size(dev_bytes);
