+++ library/src/blas1/rocblas_nrm2.hpp
@@ -5,9 +5,8 @@
 
 #include "fetch_template.h"
 #include "handle.h"
-#include "reduction_strided_batched2.h"
+#include "reduction_strided_batched.h"
 #include "rocblas.h"
-
 
 template <class To>
 struct rocblas_fetch_nrm2
@@ -28,11 +27,15 @@ struct rocblas_finalize_nrm2
     }
 };
 
-
 // allocate workspace inside this API
 template <rocblas_int NB, typename Ti, typename To>
-rocblas_status rocblas_nrm2_template(
-    rocblas_handle handle, rocblas_int n, const Ti* x, rocblas_int shiftx, rocblas_int incx, To* workspace, To* result)
+rocblas_status rocblas_nrm2_template(rocblas_handle handle,
+                                     rocblas_int    n,
+                                     const Ti*      x,
+                                     rocblas_int    shiftx,
+                                     rocblas_int    incx,
+                                     To*            workspace,
+                                     To*            result)
 {
     // Quick return if possible.
     if(n <= 0 || incx <= 0)
@@ -46,9 +49,10 @@ rocblas_status rocblas_nrm2_template(
         return rocblas_status_success;
     }
 
-    return rocblas_reduction_strided_batched_kernel<NB, Ti,
-                                    rocblas_fetch_nrm2<To>,
-                                    rocblas_reduce_sum,
-                                    rocblas_finalize_nrm2>(
+    return rocblas_reduction_strided_batched_kernel<NB,
+                                                    Ti,
+                                                    rocblas_fetch_nrm2<To>,
+                                                    rocblas_reduce_sum,
+                                                    rocblas_finalize_nrm2>(
         handle, n, x, shiftx, incx, 0, 1, workspace, result);
 }
