+++ library/src/blas1/rocblas_nrm2_batched.cpp
@@ -26,8 +26,8 @@ namespace
                                              rocblas_int     n,
                                              const Ti* const x[],
                                              rocblas_int     incx,
-                                             To*             results,
-                                             rocblas_int     batch_count)
+                                             rocblas_int     batch_count,
+                                             To*             results)
     {
         if(!handle)
             return rocblas_status_invalid_handle;
@@ -57,8 +57,7 @@ namespace
         if(batch_count < 0)
             return rocblas_status_invalid_size;
 
-        size_t dev_bytes
-            = rocblas_reduction_kernel_workspace_size<NB>(n, batch_count, results);
+        size_t dev_bytes = rocblas_reduction_kernel_workspace_size<NB>(n, batch_count, results);
 
         if(handle->is_device_memory_size_query())
             return handle->set_optimal_device_memory_size(dev_bytes);
@@ -83,44 +82,44 @@ rocblas_status rocblas_snrm2_batched(rocblas_handle     handle,
                                      rocblas_int        n,
                                      const float* const x[],
                                      rocblas_int        incx,
-                                     float*             results,
-                                     rocblas_int        batch_count)
+                                     rocblas_int        batch_count,
+                                     float*             results)
 {
     constexpr rocblas_int NB = 512;
-    return rocblas_nrm2_batched_impl<NB>(handle, n, x, incx, results, batch_count);
+    return rocblas_nrm2_batched_impl<NB>(handle, n, x, incx, batch_count, results);
 }
 
 rocblas_status rocblas_dnrm2_batched(rocblas_handle      handle,
                                      rocblas_int         n,
                                      const double* const x[],
                                      rocblas_int         incx,
-                                     double*             results,
-                                     rocblas_int         batch_count)
+                                     rocblas_int         batch_count,
+                                     double*             results)
 {
     constexpr rocblas_int NB = 512;
-    return rocblas_nrm2_batched_impl<NB>(handle, n, x, incx, results, batch_count);
+    return rocblas_nrm2_batched_impl<NB>(handle, n, x, incx, batch_count, results);
 }
 
 rocblas_status rocblas_scnrm2_batched(rocblas_handle                     handle,
                                       rocblas_int                        n,
                                       const rocblas_float_complex* const x[],
                                       rocblas_int                        incx,
-                                      float*                             results,
-                                      rocblas_int                        batch_count)
+                                      rocblas_int                        batch_count,
+                                      float*                             results)
 {
     constexpr rocblas_int NB = 512;
-    return rocblas_nrm2_batched_impl<NB>(handle, n, x, incx, results, batch_count);
+    return rocblas_nrm2_batched_impl<NB>(handle, n, x, incx, batch_count, results);
 }
 
 rocblas_status rocblas_dznrm2_batched(rocblas_handle                      handle,
                                       rocblas_int                         n,
                                       const rocblas_double_complex* const x[],
                                       rocblas_int                         incx,
-                                      double*                             results,
-                                      rocblas_int                         batch_count)
+                                      rocblas_int                         batch_count,
+                                      double*                             results)
 {
     constexpr rocblas_int NB = 512;
-    return rocblas_nrm2_batched_impl<NB>(handle, n, x, incx, results, batch_count);
+    return rocblas_nrm2_batched_impl<NB>(handle, n, x, incx, batch_count, results);
 }
 
 } // extern "C"
